Move early hints test to HttpTest
